refactor: optimize styling w/ minor refactor
