+++ src/App.js
@@ -1,6 +1,8 @@
 import React from "react";
 import { Routes } from "./routes";
+
 function App() {
   return <Routes />;
 }
+
 export default App;
