--- src/components/Card/Card.module.css
@@ -1,14 +1,0 @@
-/* Main.module.css */
-.main {
-  width: 80vw;
-  margin-left: 280px;
-  margin-top: 70px;
-  padding: 16px;
-  height: 100vh;
-  background: var(--bg-white);
-}
-
-.cards {
-  display: flex;
-  flex-direction: row;
-}
