--- src/components/Card/CardItem/EmployeeCard/EmployeeCard.module.css
@@ -1,108 +1,0 @@
-.card__item1__first:hover {
-  box-shadow: var(--box-shadow-hover);
-}
-
-.card__content:hover {
-  box-shadow: var(--box-shadow-hover);
-}
-
-.cards {
-  display: flex;
-  flex-direction: row;
-}
-
-.card__item1__first {
-  display: flex;
-  flex-direction: row;
-  justify-content: space-between;
-  background-color: var(--bg-white);
-  border-radius: 5px;
-  padding: 16px;
-  align-items: center;
-  margin-bottom: 16px;
-  box-shadow: var(--box-shadow);
-  transition: 0.3s;
-}
-
-.card__item1__first h5 {
-  font-weight: var(--font-bold);
-  font-size:var(--font-size)
-}
-
-.card__content {
-  background-color: var(--bg-white);
-  border-radius: 5px;
-  padding: 16px;
-  box-shadow: var(--box-shadow);
-  transition: 0.3s;
-}
-
-.card__content h3 {
-  margin-top: -3px;
-  cursor: pointer;
-  font-size: 1.2rem;
-  font-weight: 500;
-}
-
-.card__item1__first__items {
-  display: flex;
-  justify-content: center;
-  text-align: center;
-  align-items: center;
-}
-
-.card__item1__first__both {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  text-align: center;
-}
-
-.card__item1__first__both h5 {
-  margin-top: 15px;
-}
-
-.card__items__number {
-  display: flex;
-  flex-direction: row;
-  text-align: center;
-  justify-content: space-between;
-  padding: 0 40px;
-  margin-top: 120px;
-  font-size: 32;
-  font-weight: var(--font-bold-600);
-}
-
-.card__item1__number {
-  font-size: 30px;
-  font-weight: var(--font-bold);
-}
-
-.card__items__number__column {
-  display: flex;
-  flex-direction: column;
-}
-
-.card__items__number__column h2 {
-  font-size: 30px;
-  font-weight: var(--font-bold);
-}
-
-.item1 {
-  display: flex;
-  flex-direction: column;
-  width: 100%;
-  margin-right: 10px;
-}
-
-.items__icon {
-  background-color: var(--primary-red);
-  width: fit-content;
-  color: white;
-  padding: 14px;
-  border-radius: 5px;
-}
-
-.text__bold__500 {
-  font-weight: var(--font-bold-500);
-}
