+++ src/feature/Dashboard/Dashboard.module.css
@@ -1,3 +1,17 @@
 .container {
   width: 100%;
 }
+
+.main {
+  width: 80vw;
+  margin-left: 280px;
+  margin-top: 70px;
+  padding: 16px;
+  height: 100vh;
+  background: var(--bg-white);
+}
+
+.wrapper {
+  display: flex;
+  flex-direction: row;
+}
