+++ src/feature/Dashboard/index.jsx
@@ -1,10 +1,11 @@
-import React from 'react';
-import Navbar from '../../components/Navbar';
-import Sidebar from '../../components/Sidebar';
-import Main from '../../components/Card';
-import styles from './Dashboard.module.css';
+import React from "react";
 
-
+import Navbar from "../../components/Navbar";
+import Sidebar from "../../components/Sidebar";
+import EmployeesCard from "../../components/cards/EmployeeCard";
+import NewEmployeesCard from "../../components/cards/NewEmployeesCard";
+import WelcomeCard from "../../components/cards/WelcomeCard";
+import styles from "./Dashboard.module.css";
 
 const Dashboard = () => {
   return (
@@ -12,7 +13,15 @@ const Dashboard = () => {
       <Navbar />
       <div className={styles.container}>
         <Sidebar />
-        <Main />
+        {/* START main area */}
+        <main className={styles.main}>
+          <div className={styles.wrapper}>
+            <EmployeesCard />
+            <NewEmployeesCard />
+            <WelcomeCard />
+          </div>
+        </main>
+        {/* END main area */}
       </div>
     </div>
   );
